# bcc-rosetta-server/src/server/openApi.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "1.4.10",
    "title": "Rosetta",
    "description": "Build Once. Integrate Your Blockchain Everywhere.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
...
    "/construction/derive": {
      "post": {
        "summary": "Derive an AccountIdentifier from a PublicKey",
        "description": "Derive returns the AccountIdentifier associated with a public key. Blockchains that require an on-chain action to create an account should not implement this method.",
        "operationId": "constructionDerive",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionDeriveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionDeriveResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/preprocess": {
      "post": {
        "summary": "Create a Request to Fetch Metadata",
        "description": "Preprocess is called prior to `/construction/payloads` to construct a request for any metadata that is needed for transaction construction given (i.e. account nonce). The `options` object returned from this endpoint will be sent to the `/construction/metadata` endpoint UNMODIFIED by the caller (in an offline execution environment). If your Construction API implementation has configuration options, they MUST be specified in the `/construction/preprocess` request (in the `metadata` field).",
        "operationId": "constructionPreprocess",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionPreprocessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionPreprocessResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/metadata": {
      "post": {
        "summary": "Get Metadata for Transaction Construction",
        "description": "Get any information required to construct a transaction for a specific network. Metadata returned here could be a recent hash to use, an account sequence number, or even arbitrary chain state. The request used when calling this endpoint is created by calling `/construction/preprocess` in an offline environment. You should NEVER assume that the request sent to this endpoint will be created by the caller or populated with any custom parameters. This must occur in `/construction/preprocess`. It is important to clarify that this endpoint should not pre-construct any transactions for the client (this should happen in `/construction/payloads`). This endpoint is left purposely unstructured because of the wide scope of metadata that could be required.",
        "operationId": "constructionMetadata",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionMetadataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionMetadataResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/payloads": {
      "post": {
        "summary": "Generate an Unsigned Transaction and Signing Payloads",
        "description": "Payloads is called with an array of operations and the response from `/construction/metadata`. It returns an unsigned transaction blob and a collection of payloads that must be signed by particular AccountIdentifiers using a certain SignatureType. The array of operations provided in transaction construction often times can not specify all \"effects\" of a transaction (consider invoked transactions in Ethereum). However, they can deterministically specify the \"intent\" of the transaction, which is sufficient for construction. For this reason, parsing the corresponding transaction in the Data API (when it lands on chain) will contain a superset of whatever operations were provided during construction.",
        "operationId": "constructionPayloads",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionPayloadsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionPayloadsResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/combine": {
      "post": {
        "summary": "Create Network Transaction from Signatures",
        "description": "Combine creates a network-specific transaction from an unsigned transaction and an array of provided signatures. The signed transaction returned from this method will be sent to the `/construction/submit` endpoint by the caller.",
        "operationId": "constructionCombine",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionCombineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionCombineResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/parse": {
      "post": {
        "summary": "Parse a Transaction",
        "description": "Parse is called on both unsigned and signed transactions to understand the intent of the formulated transaction. This is run as a sanity check before signing (after `/construction/payloads`) and before broadcast (after `/construction/combine`). ",
        "operationId": "constructionParse",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionParseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionParseResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/hash": {
      "post": {
        "summary": "Get the Hash of a Signed Transaction",
        "description": "TransactionHash returns the network-specific transaction hash for a signed transaction.",
        "operationId": "constructionHash",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionHashRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionIdentifierResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/submit": {
      "post": {
        "summary": "Submit a Signed Transaction",
        "description": "Submit a pre-signed transaction to the node. This call should not block on the transaction being included in a block. Rather, it should return immediately with an indication of whether or not the transaction was included in the mempool. The transaction submission response should only return a 200 status if the submitted transaction could be included in the mempool. Otherwise, it should return an error.",
        "operationId": "constructionSubmit",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionIdentifierResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConstructionMetadataRequest": {
        "description": "A ConstructionMetadataRequest is utilized to get information required to construct a transaction. The Options object used to specify which metadata to return is left purposely unstructured to allow flexibility for implementers. Options is not required in the case that there is network-wide metadata of interest. Optionally, the request can also include an array of PublicKeys associated with the AccountIdentifiers returned in ConstructionPreprocessResponse.",
        "type": "object",
        "required": ["network_identifier", "options"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "options": {
            "description": "Some blockchains require different metadata for different types of transaction construction (ex: delegation versus a transfer). Instead of requiring a blockchain node to return all possible types of metadata for construction (which may require multiple node fetches), the client can populate an options object to limit the metadata returned to only the subset required.",
            "type": "object",
            "required": ["relative_ttl", "transaction_size"],
            "properties": {
              "relative_ttl": { "type": "number" },
              "transaction_size": { "type": "number" }
            }
          },
          "public_keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicKey"
            }
          }
        }
      },
      "ConstructionMetadataResponse": {
        "description": "The ConstructionMetadataResponse returns network-specific metadata used for transaction construction. Optionally, the implementer can return the suggested fee associated with the transaction being constructed. The caller may use this info to adjust the intent of the transaction or to create a transaction with a different account that can pay the suggested fee. Suggested fee is an array in case fee payment must occur in multiple currencies.",
        "type": "object",
        "required": ["metadata"],
        "properties": {
          "metadata": {
            "type": "object",
            "properties": {
              "ttl": { "type": "string" }
            },
            "example": {
              "account_sequence": 23,
              "recent_block_hash": "0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5"
            }
          },
          "suggested_fee": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Amount"
            }
          }
        }
      },
      "ConstructionDeriveRequest": {
        "description": "ConstructionDeriveRequest is passed to the `/construction/derive` endpoint. Network is provided in the request because some blockchains have different address formats for different networks. Metadata is provided in the request because some blockchains allow for multiple address types (i.e. different address for validators vs normal accounts).",
        "type": "object",
        "required": ["network_identifier", "public_key"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "public_key": {
            "$ref": "#/components/schemas/PublicKey"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "staking_credential": {
                "$ref": "#/components/schemas/PublicKey"
              },
              "address_type": {
                "$ref": "#/components/schemas/AddressType"
              }
            }
          }
        }
      },
      "ConstructionDeriveResponse": {
        "description": "ConstructionDeriveResponse is returned by the `/construction/derive` endpoint.",
        "type": "object",
        "required": ["account_identifier"],
        "properties": {
          "address": {
            "type": "string",
            "description": "[DEPRECATED by `account_identifier` in `v1.4.4`] Address in network-specific format."
          },
          "account_identifier": {
            "$ref": "#/components/schemas/AccountIdentifier"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "ConstructionPreprocessRequest": {
        "description": "ConstructionPreprocessRequest is passed to the `/construction/preprocess` endpoint so that a Rosetta implementation can determine which metadata it needs to request for construction. Metadata provided in this object should NEVER be a product of live data (i.e. the caller must follow some network-specific data fetching strategy outside of the Construction API to populate required Metadata). If live data is required for construction, it MUST be fetched in the call to `/construction/metadata`. The caller can provide a max fee they are willing to pay for a transaction. This is an array in the case fees must be paid in multiple currencies. The caller can also provide a suggested fee multiplier to indicate that the suggested fee should be scaled. This may be used to set higher fees for urgent transactions or to pay lower fees when there is less urgency. It is assumed that providing a very low multiplier (like 0.0001) will never lead to a transaction being created with a fee less than the minimum network fee (if applicable). In the case that the caller provides both a max fee and a suggested fee multiplier, the max fee will set an upper bound on the suggested fee (regardless of the multiplier provided).",
        "type": "object",
        "required": ["network_identifier", "operations"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Operation"
            }
          },
          "metadata": {
            "type": "object",
            "required": ["relative_ttl"],
            "properties": {
              "relative_ttl": { "type": "number" }
            }
          },
          "max_fee": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Amount"
            }
          },
          "suggested_fee_multiplier": {
            "type": "number",
            "format": "double",
            "minimum": 0
          }
        }
      },
      "ConstructionPreprocessResponse": {
        "description": "ConstructionPreprocessResponse contains `options` that will be sent unmodified to `/construction/metadata`. If it is not necessary to make a request to `/construction/metadata`, `options` should be omitted.  Some blockchains require the PublicKey of particular AccountIdentifiers to construct a valid transaction. To fetch these PublicKeys, populate `required_public_keys` with the AccountIdentifiers associated with the desired PublicKeys. If it is not necessary to retrieve any PublicKeys for construction, `required_public_keys` should be omitted.",
        "type": "object",
        "properties": {
          "options": {
            "type": "object",
            "description": "The options that will be sent directly to `/construction/metadata` by the caller.",
            "required": ["relative_ttl", "transaction_size"],
            "properties": {
              "relative_ttl": { "type": "number" },
              "transaction_size": { "type": "number" }
            }
          },
          "required_public_keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountIdentifier"
            }
          }
        }
      },
...
      "Error": {
        "description": "Instead of utilizing HTTP status codes to describe node errors (which often do not have a good analog), rich errors are returned using this object. Both the code and message fields can be individually used to correctly identify an error. Implementations MUST use unique values for both fields.",
        "type": "object",
        "required": ["code", "message", "retriable"],
        "properties": {
          "code": {
            "description": "Code is a network-specific error code. If desired, this code can be equivalent to an HTTP status code.",
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "example": 12
          },
          "message": {
            "description": "Message is a network-specific error message. The message MUST NOT change for a given code. In particular, this means that any contextual information should be included in the details field.",
            "type": "string",
            "example": "Invalid account format"
          },
          "description": {
            "description": "Description allows the implementer to optionally provide additional information about an error. In many cases, the content of this field will be a copy-and-paste from existing developer documentation. Description can ONLY be populated with generic information about a particular type of error. It MUST NOT be populated with information about a particular instantiation of an error (use `details` for this). Whereas the content of Error.Message should stay stable across releases, the content of Error.Description will likely change across releases (as implementers improve error documentation). For this reason, the content in this field is not part of any type assertion (unlike Error.Message).",
            "type": "string",
            "example": "This error is returned when the requested AccountIdentifier is improperly formatted."
          },
          "retriable": {
            "description": "An error is retriable if the same request may succeed if submitted again.",
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "required": ["message"],
            "properties": {
              "message": { "type": "string" }
            },
            "description": "Often times it is useful to return context specific to the request that caused the error (i.e. a sample of the stack trace or impacted account) in addition to the standard error message.",
            "example": {
              "address": "0x1dcc4de8dec75d7aab85b567b6",
              "error": "not base64"
            }
          }
        }
      }
    }
  }
}